scripts/generate_readme/requirements.txt
python-magic
datefinder
jinja2